Deprecated old RiveFile utils. Converted Layout, BlendModes and StateMachine UIKit examples. Commented out unused examples and old RiveView related things. Fixed SimpleAnimation presentation issues.
