+++ Example-iOS/Source/lib/utility.swift
@@ -9,6 +9,7 @@
 import Foundation
 import RiveRuntime
 
+@available(*, deprecated, message: "Use method in RiveFile+Extensions instead")
 func getBytes(resourceName: String, resourceExt: String=".riv") -> [UInt8] {
     guard let url = Bundle.main.url(forResource: resourceName, withExtension: resourceExt) else {
         fatalError("Failed to locate \(resourceName) in bundle.")
@@ -21,7 +22,7 @@ func getBytes(resourceName: String, resourceExt: String=".riv") -> [UInt8] {
     return [UInt8](data)
 }
 
-
+@available(*, deprecated, message: "Use convenience init in RiveFile+Extensions instead")
 func getRiveFile(resourceName: String, resourceExt: String=".riv") throws -> RiveFile{
     let byteArray = getBytes(resourceName: resourceName, resourceExt: resourceExt)
     return try RiveFile(byteArray: byteArray)
